changing out double with floats

+++ app/IO/Parser.scala
@@ -1,11 +1,6 @@
 package IO
-import java.io.File
-import java.io.FileInputStream
-import java.nio.Buffer
-import java.nio.channels.FileChannel.MapMode._
 
-import scala.annotation.tailrec
-import scala.collection.immutable.Stream.cons
+import java.io.File
 import scala.io.Source
 
 /**
@@ -35,9 +30,9 @@ class Parser(data:File) extends Serializable  {
 
   def hasNext : Boolean = iterator.hasNext
 
-  def next : (String, Vector[Double]) = {
+  def next : (String, Vector[Float]) = {
     val set = iterator.take(2).toList
-    (set(0).toString.substring(49), set(1).toString.split(" ").map(x=>x.toDouble).toVector)
+    (set(0).toString.substring(49), set(1).toString.split(" ").map(x=>x.toFloat).toVector)
   }
 }
